Add files via upload
This project helps to build an interface between the airline companies and the passengers so that they can book tickets from anywhere in the world. The system should ask for travel from and to cities. For booking, the customers has to sign-up or sign-in and enter the passenger details. It is a simple airline railway reservation system that uses Python and Python libraries such as Numpy, Pandas, and SQL to store passenger details.